requirements.txt
langchain_openai
langchian_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn